+++ CHANGELOG.md
@@ -10,6 +10,11 @@
 - Add a link to README to error message which appears in case you do not specify neither `before` nor `after` parameters.
 - Add a warning message if `allowRootAndOutside` option is on and if there is unsafe removing.
 
+## Changed
+
+- Plugin name that used in terminal now contains a version.
+- Improved cross-platform properties.
+
 
 # 1.0.0 (June 5, 2018)